+++ src/Serialization.cs
@@ -813,7 +813,7 @@ namespace Dec
 
             if (!node.AllowReflection)
             {
-                Dbg.Err($"Couldn't find a composition method for type {fieldType}; do you need a Converter?");
+                Dbg.Err($"Couldn't find a composition method for type {valType}; do you need a Converter?");
                 return;
             }