+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Fixed
 Passing derived classes to Recorder.Write() resulted in the wrong type when deserialized.
+Misleading error messages when attempting to Record an unsupported derived class.
 
 
 ## [v0.3.0]